MS: removed back compat for 1.5.2, 1.6.2; started deprecation updates

+++ MagicSpells/src/com/nisovin/magicspells/spells/buff/FlamewalkSpell.java
@@ -92,7 +92,7 @@ public class FlamewalkSpell extends BuffSpell {
 						if (entity instanceof Player) {
 							if (entity != player && targetPlayers) {
 								if (checkPlugins) {
-									EntityDamageByEntityEvent event = new EntityDamageByEntityEvent(player, entity, DamageCause.ENTITY_ATTACK, 1);
+									EntityDamageByEntityEvent event = new EntityDamageByEntityEvent(player, entity, DamageCause.ENTITY_ATTACK, (double)1);
 									Bukkit.getServer().getPluginManager().callEvent(event);
 									if (event.isCancelled()) {
 										continue;
